fixed with thunk
