+++ src/store/index.js
@@ -1,10 +1,5 @@
 import { applyMiddleware, createStore } from "redux";
 import { rootReducer } from "./reducers/index";
-import createSagaMiddleware from "redux-saga";
-import { rootWatcher } from "../saga/index";
+import thunk from "redux-thunk";
 
-const sagaMiddleware = createSagaMiddleware();
-
-export const store = createStore(rootReducer, applyMiddleware(sagaMiddleware));
-
-sagaMiddleware.run(rootWatcher);
+export const store = createStore(rootReducer, applyMiddleware(thunk));
